Timing Pulley Update
Servo arm cutout from timing pulley made. Changes in motor_end will have
to be made to accomodate for the additional width of the servo arm/servo
assembly.

The timing pulley is now very printable, and is nearing completion. I'm
not sure of how the belt will work on this yet, as I have not played
with an actual timing belt for it.